cleanup index.py page, renamed callbacks file
